refactor(mappings): use one file per module / smart contract
