--- src/mappings/handleForgetObject.ts
@@ -1,12 +1,0 @@
-import type { CosmosMessage } from "@subql/types-cosmos";
-import type { MsgExecuteContract } from "cosmjs-types/cosmwasm/wasm/v1/tx";
-import { ObjectariumObject } from "../types";
-import { getObjectariumObjectId } from "../utils/utils";
-
-export const handleForgetObject = async (
-    msg: CosmosMessage<MsgExecuteContract>
-): Promise<void> => {
-    const objectId = getObjectariumObjectId(msg.tx.tx.events);
-
-    objectId && (await ObjectariumObject.remove(objectId));
-};
